Fix wallets at risk table
